Tests/LibWeb/Text/expected/css/keyframes-append-rule-invalidation.txt

opacity before appendRule: 0.25
opacity after appendRule: 0.75
